made Charmander pokemon

+++ src/abilities/Ability.java
@@ -3,5 +3,5 @@ package abilities;
 import pokemon.Type;
 
 public interface Ability {
-    double dealDamage(int damage, Type attackedPokemonType);
+    int dealDamage(int damage, Type attackedPokemonType);
 }
